fun tests

+++ scripts/bash/results.sh
@@ -12,10 +12,15 @@
 basepath="/home/tkolb/bsc/data/fairseq-evals"
 filename="WERdata_test.txt"
 
-rm -r ~/bsc/data/results/*
+nfiles=$(ls ~/bsc/data/results | wc -l)
+if [[ $nfiles > 0 ]]; then rm -r ~/bsc/data/results/*
 source ~/.cache/pypoetry/virtualenvs/tkolbpoetry-0grRN4_Q-py3.6/bin/activate
 python3 ../interpretresults.py "/home/tkolb/bsc/data/results" \
     "$basepath/2021-11-11/18-12-54/$filename" \
     "$basepath/2021-11-12/10-12-27/$filename" \
-    "$basepath/2021-11-12/10-34-58/$filename"
+    "$basepath/2021-11-12/10-34-58/$filename" \
+    "$basepath/2021-11-14/15-14-42/$filename" \
+    "$basepath/2021-11-14/15-23-43/$filename" \
+    "$basepath/2021-11-14/15-30-40/$filename" \
+
 deactivate
